+++ tests/test_main.py
@@ -1,5 +1,8 @@
 from argparse import Namespace
 
+import pytest
+
+from dunamai import _run_cmd
 from dunamai.__main__ import parse_args, _VERSION_PATTERN
 
 
@@ -28,3 +31,23 @@ def test__parse_args__from():
     assert parse_args(["from", "any", "--style", "semver"]).style == "semver"
     assert parse_args(["from", "any", "--latest-tag"]).latest_tag is True
     assert parse_args(["from", "subversion", "--tag-dir", "foo"]).tag_dir == "foo"
+
+    with pytest.raises(SystemExit):
+        parse_args(["from", "unknown"])
+
+
+def test__parse_args__check():
+    assert parse_args(["check", "0.1.0"]) == Namespace(
+        command="check", version="0.1.0", style="pep440"
+    )
+    assert parse_args(["check", "0.1.0", "--style", "semver"]).style == "semver"
+    assert parse_args(["check", "0.1.0", "--style", "pvp"]).style == "pvp"
+
+    with pytest.raises(SystemExit):
+        parse_args(["check", "0.1.0", "--style", "unknown"])
+
+
+def test__cli_check():
+    _run_cmd("dunamai check 0.01.0")
+    _run_cmd("dunamai check v0.1.0", codes=[1])
+    _run_cmd("dunamai check 0.01.0 --style semver", codes=[1])